fixes type check issues on build
